handle I/O exit

+++ ymir/arch/x86/asm.zig
@@ -16,7 +16,7 @@ pub inline fn inb(port: u16) u8 {
     );
 }
 
-pub inline fn inw(port: u16) u8 {
+pub inline fn inw(port: u16) u16 {
     return asm volatile (
         \\inw %[port], %[ret]
         : [ret] "={ax}" (-> u16),
